pyproject.toml
[project]
name = "business-card"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.49.1",
    "llama-cpp-python>=0.3.16",
    "ollama>=0.6.0",
    "pillow>=11.3.0",
]
